backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
motor==3.7.1
pydantic==2.13.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.27
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
python-dotenv==1.2.2